+++ packages/runtime/test/src/runtime.ts
@@ -1,5 +1,6 @@
 import { types } from "@algo-builder/web";
-import { LogicSigAccount } from "algosdk";
+import { AccountAddress, AlgoTransferParam } from "@algo-builder/web/build/types";
+import algosdk, { LogicSigAccount } from "algosdk";
 import { assert } from "chai";
 import sinon from "sinon";
 
@@ -16,7 +17,7 @@ const programName = "basic.teal";
 const minBalance = BigInt(1e7);
 
 describe("Transfer Algo Transaction", function () {
-  const amount = 1000;
+  const amount = minBalance;
   const fee = 1000;
 
   let alice: AccountStoreI;
@@ -32,9 +33,9 @@ describe("Transfer Algo Transaction", function () {
   }
 
   this.beforeEach(() => {
-    alice = new AccountStore(minBalance);
-    bob = new AccountStore(minBalance);
-    alan = new AccountStore(minBalance);
+    alice = new AccountStore(minBalance * 10n);
+    bob = new AccountStore(minBalance * 10n);
+    alan = new AccountStore(minBalance * 10n);
     runtime = new Runtime([alice, bob, alan]);
   });
 
@@ -125,6 +126,48 @@ describe("Transfer Algo Transaction", function () {
       }),
       RUNTIME_ERRORS.GENERAL.INVALID_CLOSE_REMAINDER_TO
     );
+  });
+
+  describe("Transfer algo to implicit account", function () {
+    let externalRuntimeAccount: AccountStoreI;
+    let externalAccount: algosdk.Account;
+    this.beforeEach(function () {
+      externalAccount = new AccountStore(0).account;
+
+      const transferAlgoTx: AlgoTransferParam = {
+        type: types.TransactionType.TransferAlgo,
+        sign: types.SignType.SecretKey,
+        fromAccount: alice.account,
+        toAccountAddr: externalAccount.addr,
+        amountMicroAlgos: amount,
+        payFlags: {
+          totalFee: 1000
+        }
+      };
+
+      runtime.executeTx(transferAlgoTx);
+      // query new external account in runtime.
+      externalRuntimeAccount = runtime.getAccount(externalAccount.addr);
+    });
+
+    it("Balance of toAccountAddr should updated", () => {
+      assert.equal(externalRuntimeAccount.amount, amount);
+    });
+
+    it("Can create transaction from external account", () => {
+      const transferAlgoTx: AlgoTransferParam = {
+        type: types.TransactionType.TransferAlgo,
+        sign: types.SignType.SecretKey,
+        fromAccount: externalRuntimeAccount.account,
+        toAccountAddr: alice.address,
+        amountMicroAlgos: 1000n,
+        payFlags: {
+          totalFee: 1000
+        }
+      };
+
+      assert.doesNotThrow(() => runtime.executeTx(transferAlgoTx));
+    });
   });
 });